requirements.txt
appdirs==1.4.3
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
Django==2.2.1
django-debug-toolbar==2.0
django-summernote==0.8.11.4
fake-useragent==0.1.11
idna==2.8
lxml==4.3.4
parse==1.12.0
Pillow==6.0.0
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
python-slugify==3.0.2
pytz==2019.1
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.2
tqdm==4.32.2
urllib3==1.25.3
w3lib==1.20.0
websockets==7.0
wincertstore==0.2
psycopg2